Adds a headless browser scraper for dynamic navigation sites.
Sites with links dynamically generated by JavaScript don't
get scraped by the normal scraper. This new method is a bit more
resource intensive as it uses a headless browser, but the headless
browser option allows us to scrape more dynamic websites.

This was tested manually on the docs.rust-lang.org/book site, which
I couldn't get to convert previously. Automated tests were added
both to verify the new functionality and to make sure that the
original scraper was indeed failing with JS-rendered links.